Add support for deploying packages to pub
This doesn't have any tests because I don't really know how we *would*
test it. All it really does is invoke the pub executable, but I don't
want to test the details of the publish protocol and I don't want to
yak shave a way to mock an executable invocation.

Closes #15