Change the method for loading control data from the repository
This commit removes the config variable `debianControlData` and
introduces the `debianControlPath`.

Now, instead of passing the control data directly, the user has
to specify the path of the control file in the `debianRepo` to
load the same.

By default, the task tries reading the data from a file named
`control` in the `debianRepo`.

Signed-off-by: Shubhanshu Saxena <shubhanshu.e01@gmail.com>